# public/cap-actions-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Capital Bridge Solutions - Cap AI Actions",
    "description": "API endpoints for Cap, the AI-powered DSCR loan advisor in ChatGPT. Provides property analysis, market data, DSCR calculations, and call scheduling for real estate investors.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://www.capitalbridgesolutions.com",
      "description": "Production API"
    }
  ],
  "paths": {
...
    "/api/chatgpt/analyze-property": {
      "post": {
        "operationId": "analyzeProperty",
        "summary": "Analyze investment property metrics",
        "description": "Provides comprehensive analysis of a property including DSCR calculation, cash flow projections, and approval likelihood. Use this when investor wants to evaluate a specific property.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["purchase_price", "monthly_rent"],
                "properties": {
                  "address": {
                    "type": "string",
                    "description": "Property address"
                  },
                  "city": {
                    "type": "string",
                    "description": "City name"
                  },
                  "state": {
                    "type": "string",
                    "default": "CA",
                    "description": "State abbreviation"
                  },
                  "purchase_price": {
                    "type": "number",
                    "description": "Purchase price or current value in dollars"
                  },
                  "monthly_rent": {
                    "type": "number",
                    "description": "Monthly rental income in dollars"
                  },
                  "down_payment_percent": {
                    "type": "number",
                    "default": 25,
                    "description": "Down payment as percentage (e.g., 25 for 25%)"
                  },
                  "property_type": {
                    "type": "string",
                    "default": "Single Family",
                    "description": "Property type (e.g., Single Family, Multi-Family, Condo)"
                  },
                  "credit_score": {
                    "type": "number",
                    "description": "Credit score for rate estimation"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Property analysis results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "property": {
                      "type": "object"
                    },
                    "loan_details": {
                      "type": "object"
                    },
                    "monthly_breakdown": {
                      "type": "object"
                    },
                    "metrics": {
                      "type": "object",
                      "properties": {
                        "dscr": {
                          "type": "number"
                        },
                        "monthly_cash_flow": {
                          "type": "number"
                        },
                        "cash_on_cash_return": {
                          "type": "number"
                        }
                      }
                    },
                    "approval": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chatgpt/calculate-dscr": {
      "post": {
        "operationId": "calculateDSCR",
        "summary": "Calculate DSCR ratio and cash flow",
        "description": "Interactive DSCR calculator that computes debt service coverage ratio, monthly cash flow, and provides approval likelihood. Use this when investor wants to run different scenarios or understand DSCR calculations.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["monthly_rent"],
                "properties": {
                  "monthly_rent": {
                    "type": "number",
                    "description": "Monthly rental income in dollars"
                  },
                  "purchase_price": {
                    "type": "number",
                    "description": "Purchase price (used to calculate loan amount and expenses if not provided)"
                  },
                  "loan_amount": {
                    "type": "number",
                    "description": "Loan amount in dollars"
                  },
                  "interest_rate": {
                    "type": "number",
                    "default": 6.25,
                    "description": "Interest rate percentage (e.g., 6.25 for 6.25%)"
                  },
                  "property_tax_annual": {
                    "type": "number",
                    "description": "Annual property tax in dollars"
                  },
                  "insurance_annual": {
                    "type": "number",
                    "description": "Annual insurance premium in dollars"
                  },
                  "hoa_monthly": {
                    "type": "number",
                    "default": 0,
                    "description": "Monthly HOA fees in dollars"
                  },
                  "other_expenses_monthly": {
                    "type": "number",
                    "default": 0,
                    "description": "Other monthly expenses (maintenance, management, etc.)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DSCR calculation results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "results": {
                      "type": "object",
                      "properties": {
                        "dscr": {
                          "type": "number"
                        },
                        "dscr_rating": {
                          "type": "string"
                        },
                        "monthly_cash_flow": {
                          "type": "number"
                        },
                        "approval_likelihood": {
                          "type": "string"
                        }
                      }
                    },
                    "breakdown": {
                      "type": "object"
                    },
                    "recommendations": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}